Add files via upload
This is a simple user registration script in PHP and MYSQL. By using this user can easily register, filling username, email, password fields. After submitting the form, the data will be stored in a database. So that next time user can login with the login credential which are used while user registration.